Rename latest session task utilities
